requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
cassandra-driver==3.29.1
python-dotenv==1.0.1
pydantic==2.7.1
aiofiles==23.2.1
